Adding handler code for both raw data and arbitrary JSON
In building the mbed demo, I've found it necessary to build special
case wildcard handlers for data.  A number of the platforms are
returning very raw data or JSON of interesting formats.  This will also
help with processing other message types.

In the near future it would be interesting to modify this code to also
handle { key: ....  } where the object can be inspected for methods that
match the top level keys in non-array objects.